+++ crates/cli/src/command.rs
@@ -1,17 +1,8 @@
-mod build;
-pub use build::build;
-
-mod keypair;
-pub use keypair::{keypair, KeyPairCommand};
-
 mod fuzz;
 pub use fuzz::{fuzz, FuzzCommand};
 
-mod test;
-pub use test::test;
-
 mod init;
-pub use init::{init, TestsType};
+pub use init::init;
 
 mod clean;
 pub use clean::clean;