+++ crates/cli/src/command/init.rs
@@ -1,5 +1,4 @@
 use anyhow::{bail, Error};
-use clap::ValueEnum;
 use fehler::throws;
 use trident_client::___private::TestGenerator;
 
@@ -7,15 +6,8 @@ use crate::_discover;
 
 pub const ANCHOR_TOML: &str = "Anchor.toml";
 
-#[derive(ValueEnum, Clone)]
-pub enum TestsType {
-    Both,
-    Fuzz,
-    Poc,
-}
-
 #[throws]
-pub async fn init(tests_type: TestsType) {
+pub async fn init() {
     // look for Anchor.toml
     let root = if let Some(r) = _discover(ANCHOR_TOML)? {
         r
@@ -25,15 +17,5 @@ pub async fn init(tests_type: TestsType) {
 
     let mut generator: TestGenerator = TestGenerator::new_with_root(root);
 
-    match tests_type {
-        TestsType::Poc => {
-            generator.generate_poc().await?;
-        }
-        TestsType::Both => {
-            generator.generate_both().await?;
-        }
-        TestsType::Fuzz => {
-            generator.generate_fuzz().await?;
-        }
-    };
+    generator.generate_fuzz().await?;
 }