+++ crates/test/tests/expand/with_root.rs
@@ -1,3 +1,0 @@
-#[trident_test::trident_test(root = "i_am_root")]
-async fn test_with_defined_root() {
-}